requirements.txt
Django==3.0.2

pytz==2019.3

httplib2==0.16.0
feedparser==5.2.1
Markdown==3.0.1 # pyup: <3.1

psycopg2-binary==2.8.4
versiontools==1.9.1
pep8==1.7.1
mccabe==0.6.1

pyflakes==2.1.1
pycodestyle==2.5.0

enum34==1.1.6
configparser==4.0.2
flake8==3.7.9

cssselect==1.1.0
lxml==4.4.2
sure==1.4.11

rdflib==4.2.2

coverage==5.0.3

requests==2.22.0

logilab-common==1.5.2

pylint==2.4.4
six==1.14.0
opml==0.5

Pillow==7.0.0

pytz==2019.3

python-dateutil==2.8.1
anyjson==0.3.3
billiard==3.6.1.0

vine==1.3.0
amqp==2.5.2

amqplib==1.0.2

pyparsing==2.4.6

kombu==4.6.7

celery==4.4.0
statsd==3.3.0

gunicorn==20.0.4

pycryptodome==3.9.4
ecdsa==0.15

sqlparse==0.3.0

astroid==2.3.3

html2text==2020.1.16

boto==2.49.0
raven==6.10.0
SPARQLWrapper==1.8.5
argparse==1.4.0
backports-abc==0.5
backports.ssl-match-hostname==3.7.0.1
certifi==2019.11.28

decorator==4.4.1

colorama==0.4.3

funcsigs==1.0.2
html5lib==1.0.1
ipython-genutils==0.2.0
isodate==0.6.0
keepalive==0.5

mock==3.0.5

lazy-object-proxy==1.4.3

nose==1.3.7

pbr==5.4.4

path.py==12.4.0

pickleshare==0.7.5
ptyprocess==0.6.0
rcssmin==1.0.6
rjsmin==1.0.12
simplegeneric==0.8.1
singledispatch==3.4.0.3
traitlets==4.3.3
wrapt==1.11.2

django-appconf==1.0.3

django-bootstrap3==12.0.3

django-compressor==2.4

django-bootstrap-form==3.4

django-waffle==0.19.0

django-debug-toolbar==2.1

django-jenkins==0.110.0
django-smoketest==1.1.0

django-guardian==2.1.0

easy-thumbnails==2.7

django-userena==2.0.1
django-picklefield==2.0
django-statsd-mozilla==0.4.0

django-markwhat==1.6.2

django-storages==1.8

django-cacheds3storage==0.2.1

thraxilsettings==3.1.3

honeycomb-beeline==2.11.3
